[server] adjust the configuration of go PPROF service
