+++ server/cmd/server/main.go
@@ -20,7 +20,6 @@ import (
 	"flag"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"os"
 	"os/signal"
 	"runtime"
@@ -32,13 +31,14 @@ import (
 	"github.com/deepflowio/deepflow/server/controller/controller"
 	"github.com/deepflowio/deepflow/server/controller/report"
 	"github.com/deepflowio/deepflow/server/controller/trisolaris/utils"
+	"github.com/deepflowio/deepflow/server/ingester/droplet/profiler"
 	"github.com/deepflowio/deepflow/server/ingester/ingester"
+	"github.com/deepflowio/deepflow/server/ingester/ingesterctl"
+	"github.com/deepflowio/deepflow/server/libs/debug"
 	"github.com/deepflowio/deepflow/server/libs/logger"
 	"github.com/deepflowio/deepflow/server/querier/querier"
 
 	logging "github.com/op/go-logging"
-	"github.com/pyroscope-io/client/pyroscope"
-	yaml "gopkg.in/yaml.v2"
 )
 
 func execName() string {
@@ -48,53 +48,15 @@ func execName() string {
 
 var log = logging.MustGetLogger(execName())
 
+const (
+	PROFILER_PORT = 9526
+)
+
 var flagSet = flag.NewFlagSet(os.Args[0], flag.ExitOnError)
 var configPath = flagSet.String("f", "/etc/server.yaml", "Specify config file location")
 var version = flagSet.Bool("v", false, "Display the version")
 
 var Branch, RevCount, Revision, CommitDate, goVersion, CompileTime string
-
-type Config struct {
-	LogFile           string            `default:"/var/log/deepflow/server.log" yaml:"log-file"`
-	LogLevel          string            `default:"info" yaml:"log-level"`
-	ContinuousProfile ContinuousProfile `yaml:"continuous-profile"`
-}
-
-type ContinuousProfile struct {
-	Enabled       bool     `yaml:"enabled"`
-	ServerAddress string   `yaml:"server-addr"`
-	ProfileTypes  []string `yaml:"profile-types"` // pyroscope.ProfileType
-	MutexRate     int      `yaml:"mutex-rate"`    // valid when ProfileTypes contains 'mutex_count' or 'mutex_duration'
-	BlockRate     int      `yaml:"block-rate"`    // valid when ProfileTypes contains 'block_count' or 'block_duration'
-	LogEnabled    bool     `yaml:"log-enabled"`   // logging enabled
-}
-
-func loadConfig(path string) *Config {
-	config := &Config{
-		LogFile:  "/var/log/deepflow/server.log",
-		LogLevel: "info",
-		ContinuousProfile: ContinuousProfile{
-			Enabled:       false,
-			ServerAddress: "http://deepflow-agent/api/v1/profile",
-			ProfileTypes:  []string{"cpu", "inuse_objects", "alloc_objects", "inuse_space", "alloc_space"},
-			MutexRate:     5,
-			BlockRate:     5,
-			LogEnabled:    true,
-		},
-	}
-	configBytes, err := ioutil.ReadFile(path)
-	if err != nil {
-		fmt.Printf("Read config file path: %s, error: %s", err, path)
-		os.Exit(1)
-	}
-
-	if err = yaml.Unmarshal(configBytes, &config); err != nil {
-		fmt.Printf("Unmarshal yaml(%s) error: %s", path, err)
-		os.Exit(1)
-	}
-
-	return config
-}
 
 func main() {
 	flagSet.Parse(os.Args[1:])
@@ -118,6 +80,20 @@ func main() {
 	logging.SetLevel(logLevel, "")
 
 	log.Infof("deepflow-server config: %+v", *cfg)
+
+	debug.SetIpAndPort(ingesterctl.DEBUG_LISTEN_IP, ingesterctl.DEBUG_LISTEN_PORT)
+	debug.NewLogLevelControl()
+	profiler := profiler.NewProfiler(PROFILER_PORT)
+	if cfg.Profiler {
+		runtime.SetMutexProfileFraction(1)
+		runtime.SetBlockProfileRate(1)
+		profiler.Start()
+	}
+
+	if cfg.MaxCPUs > 0 {
+		runtime.GOMAXPROCS(cfg.MaxCPUs)
+	}
+
 	startContinuousProfile(&cfg.ContinuousProfile)
 
 	ctx, cancel := utils.NewWaitGroupCtx()
@@ -152,52 +128,4 @@ func main() {
 		}(closer)
 	}
 	wg.Wait()
-}
-
-func startContinuousProfile(config *ContinuousProfile) {
-	if !config.Enabled {
-		return
-	}
-	profileTypes := []pyroscope.ProfileType{}
-	hasMutexProfile, hasBlockProfile := false, false
-	for _, t := range config.ProfileTypes {
-		switch pyroscope.ProfileType(t) {
-		case pyroscope.ProfileCPU,
-			pyroscope.ProfileInuseObjects,
-			pyroscope.ProfileAllocObjects,
-			pyroscope.ProfileInuseSpace,
-			pyroscope.ProfileAllocSpace,
-			pyroscope.ProfileGoroutines:
-			profileTypes = append(profileTypes, pyroscope.ProfileType(t))
-		case pyroscope.ProfileMutexCount, pyroscope.ProfileMutexDuration:
-			hasMutexProfile = true
-			profileTypes = append(profileTypes, pyroscope.ProfileType(t))
-		case pyroscope.ProfileBlockCount, pyroscope.ProfileBlockDuration:
-			hasBlockProfile = true
-			profileTypes = append(profileTypes, pyroscope.ProfileType(t))
-		}
-	}
-
-	if hasMutexProfile {
-		runtime.SetMutexProfileFraction(config.MutexRate)
-	}
-	if hasBlockProfile {
-		runtime.SetBlockProfileRate(config.BlockRate)
-	}
-	logger := log
-	if !config.LogEnabled {
-		logger = nil
-	}
-
-	pyroscope.Start(pyroscope.Config{
-		ApplicationName: "deepflow-server",
-		// replace this with the address of pyroscope server
-		ServerAddress: config.ServerAddress,
-		// you can disable logging by setting this to nil
-		Logger: logger,
-		// you can provide static tags via a map:
-		Tags:         map[string]string{"hostname": os.Getenv("K8S_NODE_NAME_FOR_DEEPFLOW")},
-		ProfileTypes: profileTypes,
-	})
-
 }
